// site/src/models/StatisticsDTO.ts
export interface StatisticsDTO {
    health: number;
    power: number;
    powerType?: string;
    movementSpeed?: number;
    hitChance?: number;
    expertise?: number;
    rangedAttackPower?: number;
    strength: number;
    agility: number;
    intellect: number;
    stamina: number;
    meleeCrit: number;
    meleeHaste: number;
    mastery: number;
    bonusArmor: number;
    attackPower: number;
    mainHandDamageMin: number;
    mainHandDamageMax: number;
    mainHandSpeed: number;
    mainHandDps: number;
    offHandDamageMin: number;
    offHandDamageMax: number;
    offHandSpeed: number;
    offHandDps: number;
    spellPower: number;
    spellPenetration: number;
    spellCrit: number;
    manaRegen: number;
    manaRegenCombat: number;
    armor: number;
    dodge: number;
    parry: number;
    block: number;
    rangedCrit: number;
    rangedHaste: number;
    spellHaste: number;
    spirit: number;
}